# Add verbose logs for node/plugin scores even ranged in low levels