<!--  Thanks for sending a pull request!  Here are some tips for you:

1. If this is your first time, please read our contributor guidelines: https://git.k8s.io/community/contributors/guide/first-contribution.md#your-first-contribution and developer guide https://git.k8s.io/community/contributors/devel/development.md#development-guide
2. Please label this pull request according to what type of issue you are addressing, especially if this is a release targeted pull request. For reference on required PR/issue labels, read here:
https://git.k8s.io/community/contributors/devel/sig-release/release.md#issuepr-kind-label
3. Ensure you have added or ran the appropriate tests for your PR: https://git.k8s.io/community/contributors/devel/sig-testing/testing.md
4. If you want *faster* PR reviews, read how: https://git.k8s.io/community/contributors/guide/pull-requests.md#best-practices-for-faster-reviews
5. If the PR is unfinished, see how to mark it: https://git.k8s.io/community/contributors/guide/pull-requests.md#marking-unfinished-pull-requests
-->

#### What type of PR is this?
/kind feature

<!--
Add one of the following kinds:
/kind bug
/kind cleanup
/kind documentation
/kind feature
/kind design

Optionally add one or more of the following kinds if applicable:
/kind api-change
/kind deprecation
/kind failing-test
/kind flake
/kind regression
-->

#### What this PR does / why we need it:

No scheduling info captured if the logging level less than 10. It would be a little help to find what happend if the top N highest node/plugin scores dumped at a lower level.

@damemi made some improvement in #99411 , however it is found ~50% performace drop in large cluster. 

As @Huang-Wei suggests in  #101820，we use a dynamical logging strategy to dump depending on the logging level:
- for logging level ranged in [4, 6), dump the scores for topM nodes/plugins. And make it as performant as possible.
- for logging level ranged in [6, 10), dump the scores for topN nodes, and probably show all plugins
- for logging level >= 10, dump the scores for all nodes/plugins.


The output looks like this with -v=4:

```
I0728 15:09:41.627395   98057 scheduling_queue.go:872] "About to try and schedule pod" pod="scheduler-extenderbd2630b2-df6f-4a45-abe0-7e80e26ac76d/extender-test-pod"
I0728 15:09:41.627418   98057 scheduler.go:516] "Attempting to schedule pod" pod="scheduler-extenderbd2630b2-df6f-4a45-abe0-7e80e26ac76d/extender-test-pod"
E0728 15:09:41.627519   98057 plugin.go:138] "getting namespace, assuming empty set of namespace labels" err="namespace \"scheduler-extenderbd2630b2-df6f-4a45-abe0-7e80e26ac76d\" not found" namespace="scheduler-extenderbd2630b2-df6f-4a45-
abe0-7e80e26ac76d"
E0728 15:09:41.630656   98057 plugin.go:138] "getting namespace, assuming empty set of namespace labels" err="namespace \"scheduler-extenderbd2630b2-df6f-4a45-abe0-7e80e26ac76d\" not found" namespace="scheduler-extenderbd2630b2-df6f-4a45-
abe0-7e80e26ac76d"
I0728 15:09:41.630850   98057 generic_scheduler.go:597] "Top 3 plugins for pod on node" pod="scheduler-extenderbd2630b2-df6f-4a45-abe0-7e80e26ac76d/extender-test-pod" node="machine3" score=1000400 plugins=[{Name:NodePreferAvoidPods Score:
1000000} {Name:PodTopologySpread Score:200} {Name:TaintToleration Score:100}]
I0728 15:09:41.630875   98057 generic_scheduler.go:597] "Top 3 plugins for pod on node" pod="scheduler-extenderbd2630b2-df6f-4a45-abe0-7e80e26ac76d/extender-test-pod" node="machine2" score=1000400 plugins=[{Name:NodePreferAvoidPods Score:
1000000} {Name:PodTopologySpread Score:200} {Name:TaintToleration Score:100}]

```

The output looks like this with -v=6:

```
I0728 15:01:01.121359   96269 scheduler.go:516] "Attempting to schedule pod" pod="scheduler-extender87c5ecb8-b5bf-4723-b247-3905865f6a5d/extender-test-pod"
E0728 15:01:01.121444   96269 plugin.go:138] "getting namespace, assuming empty set of namespace labels" err="namespace \"scheduler-extender87c5ecb8-b5bf-4723-b247-3905865f6a5d\" not found" namespace="scheduler-extender87c5ecb8-b5bf-4723-
b247-3905865f6a5d"
E0728 15:01:01.123943   96269 plugin.go:138] "getting namespace, assuming empty set of namespace labels" err="namespace \"scheduler-extender87c5ecb8-b5bf-4723-b247-3905865f6a5d\" not found" namespace="scheduler-extender87c5ecb8-b5bf-4723-
b247-3905865f6a5d"
I0728 15:01:01.124131   96269 generic_scheduler.go:597] "Top 10 plugins for pod on node" pod="scheduler-extender87c5ecb8-b5bf-4723-b247-3905865f6a5d/extender-test-pod" node="machine2" score=1000400 plugins=[{Name:NodePreferAvoidPods Score
:1000000} {Name:PodTopologySpread Score:200} {Name:NodeResourcesBalancedAllocation Score:100} {Name:TaintToleration Score:100} {Name:InterPodAffinity Score:0} {Name:NodeResourcesLeastAllocated Score:0} {Name:NodeAffinity Score:0} {Name:Im
ageLocality Score:0}]
I0728 15:01:01.124156   96269 generic_scheduler.go:597] "Top 10 plugins for pod on node" pod="scheduler-extender87c5ecb8-b5bf-4723-b247-3905865f6a5d/extender-test-pod" node="machine3" score=1000400 plugins=[{Name:NodePreferAvoidPods Score
:1000000} {Name:PodTopologySpread Score:200} {Name:NodeResourcesBalancedAllocation Score:100} {Name:TaintToleration Score:100} {Name:InterPodAffinity Score:0} {Name:NodeResourcesLeastAllocated Score:0} {Name:NodeAffinity Score:0} {Name:Im
ageLocality Score:0}]

```


#### Which issue(s) this PR fixes:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #101820 

#### Special notes for your reviewer:
<!--
As #99913 mentioned, performance would be a problem if running on a large cluster. 
-->

Performance was evaluated and compared with the tool [k8s-sched-perf-stat](https://github.com/Huang-Wei/k8s-sched-perf-stat):

I did the performance test with the following 4 scenarios（2 branches x 2 logging levels）:

|         | v=1                 | v=4                 |
|---------|---------------------|---------------------|
| master  | perf-v1-clean.txt   | perf-v4-clean.txt   |
| feature | perf-v1-verbose.txt | perf-v4-verbose.txt |

with the command:
```
make test-integration WHAT=./test/integration/scheduler_perf KUBE_TEST_VMODULE=\"$KUBE_TEST_VMODULE\" KUBE_TEST_ARGS="-alsologtostderr=false -logtostderr=false -benchtime=1ns -bench=BenchmarkPerfScheduling/SchedulingBasic/5000Nodes/5000InitPods/1000PodsToSchedule -data-items-dir /tmp"
```

and the diff is evelauted:

```
++ bash -c 'go run *.go /Users/xiaoyang/go/src/k8s.io/kubernetes/perf-v1-clean.txt /Users/xiaoyang/go/src/k8s.io/kubernetes/perf-v1-verbose.txt'
+--------------------------+---------------------------+----------+--------+---------+---------+----------+
|          METRIC          |           GROUP           | QUANTILE |  UNIT  |   OLD   |   NEW   |   DIFF   |
+--------------------------+---------------------------+----------+--------+---------+---------+----------+
| SchedulingThroughput     | SchedulingBasic/5000Nodes | Average  | pods/s |  106.19 |  130.95 | +23.32%  |
| SchedulingThroughput     | SchedulingBasic/5000Nodes | Perc50   | pods/s |   92.11 |   98.62 | +7.07%   |
| SchedulingThroughput     | SchedulingBasic/5000Nodes | Perc90   | pods/s |  217.78 |  275.38 | +26.45%  |
| SchedulingThroughput     | SchedulingBasic/5000Nodes | Perc99   | pods/s |  229.00 |  275.38 | +20.25%  |
| scheduler_e2e_scheduling | SchedulingBasic/5000Nodes | Average  | ms     | 1202.25 | 1237.40 | +2.92%   |
| scheduler_e2e_scheduling | SchedulingBasic/5000Nodes | Perc50   | ms     |  348.70 |  786.11 | +125.44% |
| scheduler_e2e_scheduling | SchedulingBasic/5000Nodes | Perc90   | ms     | 3358.05 | 2956.64 | -11.95%  |
| scheduler_e2e_scheduling | SchedulingBasic/5000Nodes | Perc99   | ms     | 7474.71 | 6202.31 | -17.02%  |
| scheduler_pod_scheduling | SchedulingBasic/5000Nodes | Average  | ms     | 1360.00 | 1348.09 | -0.88%   |
| scheduler_pod_scheduling | SchedulingBasic/5000Nodes | Perc50   | ms     |  607.56 |  981.18 | +61.50%  |
| scheduler_pod_scheduling | SchedulingBasic/5000Nodes | Perc90   | ms     | 4041.90 | 3599.99 | -10.93%  |
| scheduler_pod_scheduling | SchedulingBasic/5000Nodes | Perc99   | ms     | 8367.33 | 6907.92 | -17.44%  |
+--------------------------+---------------------------+----------+--------+---------+---------+----------+
++ bash -c 'go run *.go /Users/xiaoyang/go/src/k8s.io/kubernetes/perf-v4-clean.txt /Users/xiaoyang/go/src/k8s.io/kubernetes/perf-v4-verbose.txt'
+--------------------------+---------------------------+----------+--------+----------+---------+----------+
|          METRIC          |           GROUP           | QUANTILE |  UNIT  |   OLD    |   NEW   |   DIFF   |
+--------------------------+---------------------------+----------+--------+----------+---------+----------+
| SchedulingThroughput     | SchedulingBasic/5000Nodes | Average  | pods/s |   105.88 |  138.51 | +30.82%  |
| SchedulingThroughput     | SchedulingBasic/5000Nodes | Perc50   | pods/s |    80.25 |   97.62 | +21.65%  |
| SchedulingThroughput     | SchedulingBasic/5000Nodes | Perc90   | pods/s |   223.13 |  275.67 | +23.55%  |
| SchedulingThroughput     | SchedulingBasic/5000Nodes | Perc99   | pods/s |   234.75 |  277.22 | +18.09%  |
| scheduler_e2e_scheduling | SchedulingBasic/5000Nodes | Average  | ms     |   715.89 |  865.39 | +20.88%  |
| scheduler_e2e_scheduling | SchedulingBasic/5000Nodes | Perc50   | ms     |    47.56 |  386.74 | +713.23% |
| scheduler_e2e_scheduling | SchedulingBasic/5000Nodes | Perc90   | ms     |  2282.13 | 2510.42 | +10.00%  |
| scheduler_e2e_scheduling | SchedulingBasic/5000Nodes | Perc99   | ms     |  5320.34 | 5314.16 | -0.12%   |
| scheduler_pod_scheduling | SchedulingBasic/5000Nodes | Average  | ms     |  4508.74 | 1922.55 | -57.36%  |
| scheduler_pod_scheduling | SchedulingBasic/5000Nodes | Perc50   | ms     |  4688.92 | 1747.25 | -62.74%  |
| scheduler_pod_scheduling | SchedulingBasic/5000Nodes | Perc90   | ms     |  8867.54 | 3840.48 | -56.69%  |
| scheduler_pod_scheduling | SchedulingBasic/5000Nodes | Perc99   | ms     | 11172.94 | 6716.56 | -39.89%  |
+--------------------------+---------------------------+----------+--------+----------+---------+----------+
++ bash -c 'go run *.go /Users/xiaoyang/go/src/k8s.io/kubernetes/perf-v1-clean.txt /Users/xiaoyang/go/src/k8s.io/kubernetes/perf-v4-clean.txt'
+--------------------------+---------------------------+----------+--------+---------+----------+----------+
|          METRIC          |           GROUP           | QUANTILE |  UNIT  |   OLD   |   NEW    |   DIFF   |
+--------------------------+---------------------------+----------+--------+---------+----------+----------+
| SchedulingThroughput     | SchedulingBasic/5000Nodes | Average  | pods/s |  106.19 |   105.88 | -0.29%   |
| SchedulingThroughput     | SchedulingBasic/5000Nodes | Perc50   | pods/s |   92.11 |    80.25 | -12.88%  |
| SchedulingThroughput     | SchedulingBasic/5000Nodes | Perc90   | pods/s |  217.78 |   223.13 | +2.46%   |
| SchedulingThroughput     | SchedulingBasic/5000Nodes | Perc99   | pods/s |  229.00 |   234.75 | +2.51%   |
| scheduler_e2e_scheduling | SchedulingBasic/5000Nodes | Average  | ms     | 1202.25 |   715.89 | -40.45%  |
| scheduler_e2e_scheduling | SchedulingBasic/5000Nodes | Perc50   | ms     |  348.70 |    47.56 | -86.36%  |
| scheduler_e2e_scheduling | SchedulingBasic/5000Nodes | Perc90   | ms     | 3358.05 |  2282.13 | -32.04%  |
| scheduler_e2e_scheduling | SchedulingBasic/5000Nodes | Perc99   | ms     | 7474.71 |  5320.34 | -28.82%  |
| scheduler_pod_scheduling | SchedulingBasic/5000Nodes | Average  | ms     | 1360.00 |  4508.74 | +231.53% |
| scheduler_pod_scheduling | SchedulingBasic/5000Nodes | Perc50   | ms     |  607.56 |  4688.92 | +671.77% |
| scheduler_pod_scheduling | SchedulingBasic/5000Nodes | Perc90   | ms     | 4041.90 |  8867.54 | +119.39% |
| scheduler_pod_scheduling | SchedulingBasic/5000Nodes | Perc99   | ms     | 8367.33 | 11172.94 | +33.53%  |
+--------------------------+---------------------------+----------+--------+---------+----------+----------+
++ bash -c 'go run *.go /Users/xiaoyang/go/src/k8s.io/kubernetes/perf-v1-verbose.txt /Users/xiaoyang/go/src/k8s.io/kubernetes/perf-v4-verbose.txt'
+--------------------------+---------------------------+----------+--------+---------+---------+---------+
|          METRIC          |           GROUP           | QUANTILE |  UNIT  |   OLD   |   NEW   |  DIFF   |
+--------------------------+---------------------------+----------+--------+---------+---------+---------+
| SchedulingThroughput     | SchedulingBasic/5000Nodes | Average  | pods/s |  130.95 |  138.51 | +5.77%  |
| SchedulingThroughput     | SchedulingBasic/5000Nodes | Perc50   | pods/s |   98.62 |   97.62 | -1.01%  |
| SchedulingThroughput     | SchedulingBasic/5000Nodes | Perc90   | pods/s |  275.38 |  275.67 | +0.11%  |
| SchedulingThroughput     | SchedulingBasic/5000Nodes | Perc99   | pods/s |  275.38 |  277.22 | +0.67%  |
| scheduler_e2e_scheduling | SchedulingBasic/5000Nodes | Average  | ms     | 1237.40 |  865.39 | -30.06% |
| scheduler_e2e_scheduling | SchedulingBasic/5000Nodes | Perc50   | ms     |  786.11 |  386.74 | -50.80% |
| scheduler_e2e_scheduling | SchedulingBasic/5000Nodes | Perc90   | ms     | 2956.64 | 2510.42 | -15.09% |
| scheduler_e2e_scheduling | SchedulingBasic/5000Nodes | Perc99   | ms     | 6202.31 | 5314.16 | -14.32% |
| scheduler_pod_scheduling | SchedulingBasic/5000Nodes | Average  | ms     | 1348.09 | 1922.55 | +42.61% |
| scheduler_pod_scheduling | SchedulingBasic/5000Nodes | Perc50   | ms     |  981.18 | 1747.25 | +78.08% |
| scheduler_pod_scheduling | SchedulingBasic/5000Nodes | Perc90   | ms     | 3599.99 | 3840.48 | +6.68%  |
| scheduler_pod_scheduling | SchedulingBasic/5000Nodes | Perc99   | ms     | 6907.92 | 6716.56 | -2.77%  |
+--------------------------+---------------------------+----------+--------+---------+---------+---------+
```

#### Does this PR introduce a user-facing change?
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".

For more information on release notes see: https://git.k8s.io/community/contributors/guide/release-notes.md
-->
```release-note
kube-scheduler now logs node and plugin scoring  even though --v<10
- socres of the top 3 plugins in the top 3 nodes are dumped if --v=4,5
- socres of all plugins in the top 6 nodes are dumped if --v=6,7,8,9
```

#### Additional documentation e.g., KEPs (Kubernetes Enhancement Proposals), usage docs, etc.:

<!--
This section can be blank if this pull request does not require a release note.

When adding links which point to resources within git repositories, like
KEPs or supporting documentation, please reference a specific commit and avoid
linking directly to the master branch. This ensures that links reference a
specific point in time, rather than a document that may change over time.

See here for guidance on getting permanent links to files: https://help.github.com/en/articles/getting-permanent-links-to-files

Please use the following format for linking documentation:
- [KEP]: <link>
- [Usage]: <link>
- [Other doc]: <link>
-->
```docs

```