@muma378 2021-07-09
> @muma378 is this ready for review? If so please move it from "draft" to ready so we know it's not still a WIP :)

very thanks for your remind, but i‘m still strugging in running e2e-test in my local laptop to check if the logs were output correctly. 
Do you have any idea to help me to get the kube-scheduler logs without building a new cluster? any suggest would be appreciate.