@Huang-Wei 2021-09-08
> /lgtm

Well, I think this made this PR merged without thorough reviewing. @wzshiming please hold off the directive `/lgtm` if it needs further discussion.

After taking a glance, I can spot several issues:

1. A consensus on the value `M` and `N`. This is not fully discussed yet.
2. Further dig into the performance result. The test runs on v=1 and v=4, but actually they're the same if it's running in the context of integration framework.
    https://github.com/kubernetes/kubernetes/blob/90851a0fb5bbb6f02b3c22715f9f91609f8ee438/test/integration/framework/master_utils.go#L131-L136
   I'd expect to see the result comparing before/after on v=4 and v=6.
3. Further review on the code, like adding UTs and some logic nits.