@damemi 2021-09-08
I didn't realize the performance tests didn't cover all the cases, sorry. We should try to at least get some measurements for that first. If there is still a big hit to performance then we should revert