@ahg-g 2021-09-08
+1 to reverting. The logic seems a bit ad hoc to me and not necessarily placed in the right pkg.

If we are serious about dumping internal scheduler state or decisions, we need to design a library for that, which includes the logic we currently have for [dumping the cache](https://github.com/kubernetes/kubernetes/tree/master/pkg/scheduler/internal/cache/debugger):