@Huang-Wei 2021-09-08
@damemi I created #104849 to revert this PR.

@muma378 Apologize for the reverting. It doesn't we don't accept the changes. It's just we want to review it thoroughly. Feel free to open a new PR, taking the comments https://github.com/kubernetes/kubernetes/pull/103515#issuecomment-914922411 and https://github.com/kubernetes/kubernetes/pull/103515#issuecomment-915421527 into consideration.