@muma378 2021-09-09
> @damemi I created #104849 to revert this PR.
> 
> @muma378 Apologize for the reverting. It doesn't we don't accept the changes. It's just we want to review it thoroughly. Feel free to open a new PR, taking the comments [#103515 (comment)](https://github.com/kubernetes/kubernetes/pull/103515#issuecomment-914922411) and [#103515 (comment)](https://github.com/kubernetes/kubernetes/pull/103515#issuecomment-915421527) into consideration.

@Huang-Wei  Never mind, it is absolutely ok with me. I feel sorry about the mistake on performance testing and will keep following the issue.

In terms of your and @ahg-g 's suggestion:

- I will do more performance test with logging levels 5/6/9, and add UTs;
- Is there anything I can do about the value choosing of `M` and `N`? (such as comparing the performace drop between different values?)
- https://github.com/kubernetes/kubernetes/pull/103515#issuecomment-915421527 sounds reasonable, however it may be a big work. Should I just open a new PR or start it in a more formal way (such as writting a design proposal)?